requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
Pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
